+++ HPXMLtoOpenStudio/resources/constructions.rb
@@ -2302,14 +2302,34 @@ module Constructions
   # @param hpxml_bldg [HPXML::Building] HPXML Building object representing an individual dwelling unit
   # @return [TODO] TODO
   def self.apply_window_skylight_shading(model, window_or_skylight, sub_surface, shading_schedules, hpxml_header, hpxml_bldg)
+    # Interior/exterior shading factors
     isf_summer = window_or_skylight.interior_shading_factor_summer
     isf_winter = window_or_skylight.interior_shading_factor_winter
     esf_summer = window_or_skylight.exterior_shading_factor_summer.nil? ? 1.0 : window_or_skylight.exterior_shading_factor_summer
     esf_winter = window_or_skylight.exterior_shading_factor_winter.nil? ? 1.0 : window_or_skylight.exterior_shading_factor_winter
-    sf_summer = isf_summer * esf_summer
-    sf_winter = isf_winter * esf_winter
-    if (sf_summer < 1.0) || (sf_winter < 1.0)
-      # Apply shading
+
+    # Insect screen factors
+    is_summer = 1.0
+    is_winter = 1.0
+    if (not window_or_skylight.insect_screen_present.nil?) && window_or_skylight.insect_screen_present
+      # Values based on "Measurement of the Solar Heat Gain Coefficient and U Value of Windows
+      # with Insect Screens" (1999) by Brunger, Dubrous, and Harrison
+      is_int_reduction = 0.15 # 15%
+      is_ext_reduction = 0.46 # 46%
+      if window_or_skylight.insect_screen_location == HPXML::LocationInterior
+        is_summer = 1.0 - is_int_reduction * window_or_skylight.insect_screen_summer_fraction_covered
+        is_winter = 1.0 - is_int_reduction * window_or_skylight.insect_screen_winter_fraction_covered
+      elsif window_or_skylight.insect_screen_location == HPXML::LocationExterior
+        is_summer = 1.0 - is_ext_reduction * window_or_skylight.insect_screen_summer_fraction_covered
+        is_winter = 1.0 - is_ext_reduction * window_or_skylight.insect_screen_winter_fraction_covered
+      end
+    end
+
+    # Total factors
+    sf_summer = isf_summer * esf_summer * is_summer
+    sf_winter = isf_winter * esf_winter * is_winter
+
+    if (sf_summer < 1.0) || (sf_winter < 1.0) # Apply shading
 
       # Determine transmittance values throughout the year
       sf_values = []